pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alphaflow"
version = "0.1.1"
description = "AlphaFlow is a Python-based, event-driven backtesting framework designed for professional-grade trading research and strategy development."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Brandon Schabell", email = "brandonschabell@gmail.com"}
]
classifiers = [
    'Development Status :: 4 - Beta',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'Intended Audience :: Financial and Insurance Industry',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Topic :: Office/Business :: Financial',
    'Topic :: Scientific/Engineering :: Information Analysis',
    'Topic :: Software Development',
]
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/brandonschabell/alphaflow"
Repository = "https://github.com/brandonschabell/alphaflow"
Issues = "https://github.com/brandonschabell/alphaflow/issues"
Changelog = "https://github.com/brandonschabell/alphaflow/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = [
  "alphaflow",
]
exclude = [
  "alphaflow/tests",
  "alphaflow/examples",
]

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
  "/docs",
  "/site",
  "*.md",
  ".gitignore",
  ".readthedocs.yaml",
  "mkdocs.yml",
]
include = [
  "/alphaflow",
  "README.md",
  "LICENSE",
  "CHANGELOG.md",
  "pyproject.toml",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "factory-boy>=3.3.0",
    "mkdocs-material>=9.6.21",
    "mkdocstrings[python]>=0.24.0",
    "mypy>=1.18.2",
    "pandas-stubs>=2.3.2.250926",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.14.0",
    "types-seaborn>=0.13.2.20250914",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "PT",  # flake8-pytest-style
    "D",   # pydocstyle (docstring linting)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "D203",  # one-blank-line-before-class (conflicts with D211)
    "D213",  # multi-line-summary-second-line (conflicts with D212)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[tool.coverage.run]
source = ["alphaflow"]
omit = [
    "alphaflow/tests/*",
    "alphaflow/examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
